requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
assemblyai==0.34.0
attrs==24.2.0
azure-core==1.31.0
azure-storage-blob==12.23.1
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
cachelib==0.9.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
dataclasses-json==0.6.7
distro==1.9.0
emoji==2.14.0
eval_type_backport==0.2.0
faiss-cpu==1.9.0
fastapi==0.115.4
filelock==3.16.1
filetype==1.2.0
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Compress==1.17
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
groq==0.11.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.1
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonify==0.5
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.13
langchain-groq==0.2.0
langchain-text-splitters==0.3.0
langdetect==1.0.9
langsmith==0.1.137
lxml==5.3.0
MarkupSafe==3.0.2
marshmallow==3.23.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
olefile==0.47
orjson==3.10.10
packaging==24.1
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
propcache==0.2.0
psutil==6.1.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pydub==0.25.1
pypdf==5.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-multipart==0.0.16
python-oxmsg==0.0.1
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.1
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.2
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.1
torch==2.5.0
tqdm==4.66.6
transformers==4.46.0
triton==3.1.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.16.5
unstructured-client==0.27.0
urllib3==2.2.3
uvicorn==0.32.0
webencodings==0.5.1
websockets==13.1
Werkzeug==3.0.6
wrapt==1.16.0
yarl==1.16.0
zstandard==0.23.0
